requirements.txt
annotated-types==0.7.0
anyio==4.2.0
certifi==2024.2.2
cffi==1.17.1
chardet==5.2.0
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
jiter==0.9.0
numpy==2.0.2
openai==1.68.0
pillow==10.2.0
protobuf==4.25.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-telegram-bot==20.8
reportlab==4.1.0
sniffio==1.3.0
sounddevice==0.5.1
tqdm==4.67.1
typing_extensions==4.12.2
